Crate doryen_rs

Structs

App

This is the game application. It handles the creation of the game window, the window events including player input events and runs the main game loop.

AppOptions

Console

This contains the data for a console (including the one displayed on the screen) and methods to draw on it.

FileLoader

This provides a common way to load files for both native and web targets

Image

An easy way to load PNG images and blit them on the console

Keys

An iterator visiting all keys in arbitrary order.

Enums

TextAlign

UpdateEvent

What is returned by the Engine::update function

Constants

CHAR_CORNER_NE

CHAR_CORNER_NW

CHAR_CORNER_SE

CHAR_CORNER_SW

CHAR_LINE_H

CHAR_LINE_V

CHAR_SUBP_DIAG

CHAR_SUBP_E

CHAR_SUBP_N

CHAR_SUBP_NE

CHAR_SUBP_NW

CHAR_SUBP_SE

CHAR_SUBP_SW

DEFAULT_CONSOLE_HEIGHT

DEFAULT_CONSOLE_WIDTH

Traits

DoryenApi

This is the complete doryen-rs API provided to you by App in Engine::update and Engine::render methods.

Engine

This is the trait you must implement to update and render your game. See App::set_engine

InputApi

Provides information about user input. Possible values for the key scancode parameter can be found in unrust/uni-app’s translate_scan_code function. Warning, there are some slight variations from one OS to another, for example the Command, F13, F14, F15 keys only exist on Mac.

Functions

color_add

color_blend

color_dist

color_mul

color_scale

Type Aliases

Color