Struct doryen_rs::AppOptions

pub struct AppOptions {
    pub console_width: u32,
    pub console_height: u32,
    pub screen_width: u32,
    pub screen_height: u32,
    pub window_title: String,
    pub font_path: String,
    pub vsync: bool,
    pub fullscreen: bool,
    pub show_cursor: bool,
    pub resizable: bool,
    pub intercept_close_request: bool,
    pub max_fps: usize,
}

Fields
console_width: u32

the console width in characters. Default is 80

console_height: u32

the console height in characters. Default is 45

screen_width: u32

the game window width in pixels

screen_height: u32

the game window height in pixels

window_title: String

the title of the game window (only in native mode)

font_path: String

the font to use. See DoryenApi::set_font_path. Default is ‘terminal_8x8.png’

vsync: bool

whether framerate are limited by the screen frequency. On web platforms, this parameter is ignored and vsync is always enabled. Default is true.

fullscreen: bool

Native only. Might not work on every platforms. Default is false.

show_cursor: bool

Whether the mouse cursor should be visible in the game window. Default is true.

resizable: bool

Whether the game window can be resized. Default is true.

intercept_close_request: bool

Intercepts clicks on the window close button. Can be checked with InputApi::close_requested Default is false (clicking on the window close button exits the game)

max_fps: usize

Limit the number of frames per second to lower CPU consumption. Use 0 for unlimited. Note that if vsync = true, frames per second cannot go over the screen refresh rate.